UI: rename / remove files
