--- ui/src/components/restqa/project-editor/RestQAProjectEditor.scss
@@ -1,9 +1,0 @@
-i.tab-header {
-  &.el-icon-success {
-    color: green;
-  }
-
-  &.el-icon-error {
-    color: red;
-  }
-}
