--- ui/src/components/restqa/project-editor/ide/RestQAProjectEditorIde.scss
@@ -1,23 +1,0 @@
-.ide {
-  /* we dont use `language-` classes anymore so thats why we need to add background and text color manually */
-  background: #2d2d2d;
-  color: #ccc;
-
-  /* you must provide font-family font-size line-height. Example: */
-  font-family: Fira code, Fira Mono, Consolas, Menlo, Courier, monospace;
-  font-size: 14px;
-  line-height: 1.5;
-  padding: 5px;
-}
-
-/* optional class for removing the outline */
-.prism-editor__textarea:focus {
-  outline: none;
-}
-
-.btn-save {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  float: right;
-  width: 100%;
-}
